Fixup

+++ enterprise/cmd/executor/release_binary.sh
@@ -7,9 +7,11 @@ set -eu
 
 # Copy uploaded binary to 'latest'.
 # TODO: Revise this. latest should probably not exist at all.
-gsutil rm -rf gs://sourcegraph-artifacts/executor/latest || true
-gsutil cp -r "gs://sourcegraph-artifacts/executor/$(git rev-parse HEAD)" gs://sourcegraph-artifacts/executor/latest
-gsutil iam ch allUsers:objectViewer gs://sourcegraph-artifacts
+if [ "${BUILDKITE_BRANCH}" = "main" ]; then
+  gsutil rm -rf gs://sourcegraph-artifacts/executor/latest || true
+  gsutil cp -r "gs://sourcegraph-artifacts/executor/$(git rev-parse HEAD)" gs://sourcegraph-artifacts/executor/latest
+  gsutil iam ch allUsers:objectViewer gs://sourcegraph-artifacts
+fi
 
 # If this is a tagged release, we want to create a directory for it.
 if [ "${EXECUTOR_IS_TAGGED_RELEASE}" = "true" ]; then
